Feat/chat be

+++ mood-tracker-be/src/main/java/com/teamc/moodtracker/dto/chat/ChatMessage.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 // 채팅방 입장 시 기존 메시지를 불러오기 위한 DTO
 public class ChatMessage {
-    private Long memberId;
+    private int memberId;
     private String message;
     private String sendTime;
 }
